requirements.txt
opencv-python>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
Flask>=3.0.0
python-dotenv>=1.2.1
dotenv>=0.9.9
google-generativeai>=0.8.5
openai>=2.9.0